requirements.txt
bcrypt==4.1.2
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.3.0
eventlet==0.35.2
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-MQTT==1.2.1
Flask-Scss==0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib_metadata==8.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
paho-mqtt==1.6.1
pymodbus==3.6.6
pyScss==1.4.0
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
urllib3==2.2.1
websocket-client==1.7.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
zipp==3.23.0
